requirements.txt
celery==5.2.7
dj-database-url==1.0.0
django==3.2.9
django-celery-beat==2.3.0
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
environs~=9.5.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dotenv~=0.21.0
python-telegram-bot==13.11
pytz
redis==4.3.4
requests
uvicorn==0.18.3
whitenoise==6.2.0
